+++ Leomard/AppStorage/UserPreferences.swift
@@ -21,8 +21,10 @@ final class UserPreferences: ObservableObject {
     @AppStorage("comment_sort_method", store: .standard) var commentSortMethod: CommentSortType = .top
     @AppStorage("check_notifs_every", store: .standard) var checkNotifsEverySeconds: Int = 60
     @AppStorage("unreadonly_when_opening_inbox", store: .standard) var unreadonlyWhenOpeningInbox: Bool = true
+    @AppStorage("profile_sort_method", store: .standard) var profileSortMethod: SortType = .new
     
     @AppStorage("experiment_x_instance_search", store: .standard) var experimentXInstanceSearch: Bool = false
     
     let sortTypes: [SortType] = [ .topHour, .topDay, .topMonth, .topYear, .hot, .active, .new, .mostComments ]
+    let profileSortTypes: [SortType] = [ .topWeek, .topMonth, .topYear, .hot, .active, .new, .mostComments, .old ]
 }