+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 0.2 (TBA)
+
+### Added
+
+- Added status indicator when sending a post (#7)
+- Added alert when sending/editing post/comment fails
+- Refreh button in the inbox
+- Replies sort method for inbox
+- Profile view sort method
+- Added status indicator when sending a reply
+
+### Changes
+
+- Post views are now a bit prettier (thanks to boscojwho on GitHub)
+
+### Bug Fixes
+
+- Fixed notification counter not updating (#9)
+- Unread message count should update, as soon as you reply to a message from the inbox
+
 ## 0.1 (2023-07-17)
 
 - Initial release